Refactor tooltip
